getAllBooks pagination
