+++ src/app/reducers/book.ts
@@ -11,6 +11,7 @@ import { Book } from '../models/book';
 
 type InitialState = {
   books: Book[];
+  count: number;
   isGetAllBooksLoading: boolean;
   isGetAllBooksDone: boolean;
   isGetAllBooksError: string;
@@ -22,6 +23,7 @@ type InitialState = {
 
 export const initialState: InitialState = {
   books: [],
+  count: 0,
   isGetAllBooksLoading: false,
   isGetAllBooksDone: false,
   isGetAllBooksError: '',
@@ -36,7 +38,7 @@ function bookReducer(state = initialState, action: BookActionTypes) {
     case GET_ALL_BOOKS_REQUEST:
       return { ...state, isGetAllBooksLoading: true };
     case GET_ALL_BOOKS_SUCCESS:
-      return { ...state, isGetAllBooksLoading: false, isGetAllBooksDone: true, books: action.payload };
+      return { ...state, isGetAllBooksLoading: false, isGetAllBooksDone: true, books: action.payload, count: action.count };
     case GET_ALL_BOOKS_FAILURE:
       return { ...state, isGetAllBooksLoading: false, isGetAllBooksDone: false, isGetAllBooksError: action.error };
