+++ src/app/sagas/book.ts
@@ -10,18 +10,19 @@ import {
   GET_BOOK_SUCCESS,
   GET_BOOK_FAILURE,
 } from '../actions/constants';
-import { GetBookRequestAction } from '../actions/types';
+import { GetAllBooksRequestAction, GetBookRequestAction } from '../actions/types';
 
-function getAllBooksAPI() {
-  return axios.get('/book');
+function getAllBooksAPI(page: number, pageSize: number) {
+  return axios.get(`/book?page=${page}&pageSize=${pageSize}`);
 }
 
-export function* getAllBooks(): SagaIterator {
+export function* getAllBooks(action: GetAllBooksRequestAction): SagaIterator {
   try {
-    const response: any = yield call(getAllBooksAPI);
+    const response: any = yield call(getAllBooksAPI, action.page, action.pageSize);
     yield put({
       type: GET_ALL_BOOKS_SUCCESS,
       payload: response.data.books,
+      count: response.data.count,
     });
   } catch (err: any) {
     yield put({
